examples/requirements.txt
azure-ai-evaluation==1.3.0